requirements.txt
pyautogen>=0.2.0
python-binance>=1.0.19
ccxt>=4.0.0
plotly>=5.0.0
pandas>=2.0.0
numpy>=1.22
python-dotenv>=1.0.0
aiohttp==3.9.0
backtrader
vectorbt
streamlit
python-binance
TA_Lib-0.4.28-cp310-cp310-win_amd64.whl
websockets
filelock